修正直接傳送 POST 要求可以繞過密碼驗證的問題

+++ delete.php
@@ -1,7 +1,8 @@
 <?php
+session_start();
 include_once 'dbconfig.php';
 
-if ($_POST['del_time']) {
+if (isset($_SESSION['authed']) && $_POST['del_time']) {
     $time = $_POST['del_time'];
     $stmt=$db_con->prepare('DELETE FROM little_diary_posts WHERE time=:time');
     $stmt->execute(array(':time'=>$time));
